Revert "Revert cinder move to plugin"
Re-apply cinder move to CSI - yes this currently is broken for GPUs on
train. However, it turns out that this does fix a slow mount where it
times out using the native plugin.

On balance, lets take the broken GPU (since we don't offer) and
investigate when it was fixed upstream since this is actively maintained
unlike the deprecated one.