Remove prefix admin
Sometimes, the organisers of a course request a couple of admin
accounts, and they provide for their names. One recent example was a
course where the organisers wanted 2 admin accounts named "jakob" and
"tomoteam".  We don't want to add the prefix "admin-" to them.

For the CLOUD Group members, we can always add the "admin-" explicitly
if we want.