Fix indentation
It is possible that the leading whitespaces in some of the JINJA
directives, like "{% if ... %}" or "{% for ... %}" are indeed adding
extra spaces in the final YAML file.

For the time being, this commit should fix the broken generated
indentations in the YAML output file.

It would be helpful to investigate and, if needed, fix the indentitation
in this JINJA template for those directives.