Allow multiple username patterns
Sometimes, the organisers requested the creation of 5 accounts
with names "trainer-1" to "trainer-5", besides the rest of usernames
with a different pattern.
Othertimes, we need to reuse the same profile for more than one course.

So, just in case something similar happens in the future, instead of
allowing only a single username, we can user multiple patterns.

They are implemented as a dictionary underneath directory "group_vars",
with format
  <prefix>: <number of users>

Then, we only need to loop over the pairs prefix:count in the Jinja2
template.

Also, use Jinja format for comments, so they do not show up in the
final output.