Allow multiple username patterns
Sometimes, the organisers requested the creation of 5 accounts
with names "trainer-1" to "trainer-5", besides the rest of usernames
with a different pattern.
Othertimes, we need to reuse the same profile for more than one course.

So, just in case something similar happens in the future, instead of
allowing only a single username, we can user multiple patterns.

They are implemented as a dictionary underneath directive "group_vars",
with format
  <prefix>: <number of users>

Then, we only need to loop over the pairs prefix:count in the Jinja2
template.
In the Jinja2 template we also check that the dictionary is not empty.
When no usernames are yet being set, but we leave the example lines,
starting with #, Jinja interprets that variable "group_vars" is defined,
but it is None, so it still attempts to loop over it and fails. So we
check it is defined AND it is not None.

We also check that the list of admin names is not empty.

Also, use Jinja format for comments, so they do not show up in the
final output.